+++ exporters/trace/instana/src/main/java/io/opencensus/exporter/trace/instana/InstanaExporterHandler.java
@@ -127,6 +127,7 @@ final class InstanaExporterHandler extends SpanExporter.Handler {
         returnToString,
         returnToString,
         returnToString,
+        returnToString,
         Functions.</*@Nullable*/ String>returnNull());
   }